chore(metrics): move each aggregator in its own file
