+++ packages/opentelemetry-metrics/src/export/aggregators/index.ts
@@ -1,5 +1,5 @@
 /*!
- * Copyright 2019, OpenTelemetry Authors
+ * Copyright 2020, OpenTelemetry Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,6 @@
  * limitations under the License.
  */
 
-export * from './BoundInstrument';
-export * from './Meter';
-export * from './Metric';
-export * from './MeterProvider';
-export * from './export/aggregators';
-export * from './export/ConsoleMetricExporter';
-export * from './export/types';
+export * from './countersum';
+export * from './observer';
+export * from './measureexact';
